feat(pass-style): extract passStyleOf to new pass-style package
