+++ packages/marshal/index.js
@@ -1,26 +1,4 @@
-export { mapIterable, filterIterable } from './src/helpers/iter-helpers.js';
-export {
-  PASS_STYLE,
-  isObject,
-  assertChecker,
-  getTag,
-  hasOwnPropertyOf,
-} from './src/helpers/passStyle-helpers.js';
-
-export { getErrorConstructor, toPassableError } from './src/helpers/error.js';
-export { getInterfaceOf } from './src/helpers/remotable.js';
-
-export {
-  nameForPassableSymbol,
-  passableSymbolForName,
-} from './src/helpers/symbol.js';
-
-export { passStyleOf, assertPassable } from './src/passStyleOf.js';
-
 export { deeplyFulfilled } from './src/deeplyFulfilled.js';
-
-export { makeTagged } from './src/makeTagged.js';
-export { Remotable, Far, ToFarFunction } from './src/make-far.js';
 
 export { QCLASS } from './src/encodeToCapData.js';
 export { makeMarshal } from './src/marshal.js';
@@ -28,14 +6,9 @@ export { stringify, parse } from './src/marshal-stringify.js';
 
 export { decodeToJustin } from './src/marshal-justin.js';
 
-export {
-  assertRecord,
-  assertCopyArray,
-  assertRemotable,
-  isRemotable,
-  isRecord,
-  isCopyArray,
-} from './src/typeGuards.js';
-
 // eslint-disable-next-line import/export
 export * from './src/types.js';
+
+// For compatibility, but importers of these should instead import these
+// directly from `@endo/pass-style` or (if applicable) `@endo/far`.
+export * from '@endo/pass-style';
