+++ packages/marshal/src/encodeToCapData.js
@@ -6,20 +6,17 @@
 // encodes to CapData, a JSON-representable data structure, and leaves it to
 // the caller (`marshal.js`) to stringify it.
 
-import { passStyleOf } from './passStyleOf.js';
-
-import { ErrorHelper } from './helpers/error.js';
-import { makeTagged } from './makeTagged.js';
 import {
+  passStyleOf,
+  isErrorLike,
+  makeTagged,
   isObject,
   getTag,
   hasOwnPropertyOf,
-} from './helpers/passStyle-helpers.js';
-import {
   assertPassableSymbol,
   nameForPassableSymbol,
   passableSymbolForName,
-} from './helpers/symbol.js';
+} from '@endo/pass-style';
 
 /** @typedef {import('./types.js').Passable} Passable */
 /** @typedef {import('./types.js').Encoding} Encoding */
@@ -249,7 +246,7 @@ export const makeEncodeToCapData = ({
     }
   };
   const encodeToCapData = passable => {
-    if (ErrorHelper.canBeValid(passable)) {
+    if (isErrorLike(passable)) {
       // We pull out this special case to accommodate errors that are not
       // valid Passables. For example, because they're not frozen.
       // The special case can only ever apply at the root, and therefore
