+++ packages/marshal/test/test-marshal-stringify.js
@@ -1,6 +1,7 @@
 import { test } from './prepare-test-env-ava.js';
 
-import { Far } from '../src/make-far.js';
+// eslint-disable-next-line import/order
+import { Far } from '@endo/pass-style';
 import { stringify, parse } from '../src/marshal-stringify.js';
 import { roundTripPairs } from './test-marshal-capdata.js';
