+++ packages/pass-style/doc/copyArray-guarantees.md
