+++ packages/pass-style/index.js
@@ -1,0 +1,35 @@
+export { mapIterable, filterIterable } from './src/iter-helpers.js';
+export {
+  PASS_STYLE,
+  isObject,
+  assertChecker,
+  getTag,
+  hasOwnPropertyOf,
+} from './src/passStyle-helpers.js';
+
+export {
+  getErrorConstructor,
+  toPassableError,
+  isErrorLike,
+} from './src/error.js';
+export { getInterfaceOf } from './src/remotable.js';
+
+export {
+  nameForPassableSymbol,
+  passableSymbolForName,
+  assertPassableSymbol,
+} from './src/symbol.js';
+
+export { passStyleOf, assertPassable } from './src/passStyleOf.js';
+
+export { makeTagged } from './src/makeTagged.js';
+export { Remotable, Far, ToFarFunction } from './src/make-far.js';
+
+export {
+  assertRecord,
+  assertCopyArray,
+  assertRemotable,
+  isRemotable,
+  isRecord,
+  isCopyArray,
+} from './src/typeGuards.js';
