+++ packages/pass-style/src/copyArray.js
