+++ packages/pass-style/src/copyRecord.js
