+++ packages/pass-style/src/internal-types.js
@@ -1,10 +1,8 @@
-/// <reference path="../extra-types.d.ts" />
-
 export {};
 
-/** @typedef {import('../types.js').Checker} Checker */
-/** @typedef {import('../types.js').PassStyle} PassStyle */
-/** @typedef {import('../types.js').PassStyleOf} PassStyleOf */
+/** @typedef {import('./types.js').Checker} Checker */
+/** @typedef {import('./types.js').PassStyle} PassStyle */
+/** @typedef {import('./types.js').PassStyleOf} PassStyleOf */
 
 /**
  * The PassStyleHelper are only used to make a `passStyleOf` function.
