+++ packages/pass-style/src/passStyle-helpers.js
@@ -1,7 +1,7 @@
 /// <reference types="ses"/>
 
-/** @typedef {import('../types.js').Checker} Checker */
-/** @typedef {import('../types.js').PassStyle} PassStyle */
+/** @typedef {import('./types.js').Checker} Checker */
+/** @typedef {import('./types.js').PassStyle} PassStyle */
 
 const { details: X, quote: q } = assert;
 const { isArray } = Array;
