+++ packages/pass-style/src/remotable.js
@@ -11,11 +11,11 @@ import {
   getTag,
 } from './passStyle-helpers.js';
 
-/** @typedef {import('../types.js').Checker} Checker */
-/** @typedef {import('../types.js').InterfaceSpec} InterfaceSpec */
-/** @typedef {import('../types.js').MarshalGetInterfaceOf} MarshalGetInterfaceOf */
+/** @typedef {import('./types.js').Checker} Checker */
+/** @typedef {import('./types.js').InterfaceSpec} InterfaceSpec */
+/** @typedef {import('./types.js').MarshalGetInterfaceOf} MarshalGetInterfaceOf */
 /** @typedef {import('./internal-types.js').PassStyleHelper} PassStyleHelper */
-/** @typedef {import('../types.js').Remotable} Remotable */
+/** @typedef {import('./types.js').Remotable} Remotable */
 
 const { details: X, Fail, quote: q } = assert;
 const { ownKeys } = Reflect;
