+++ packages/pass-style/src/typeGuards.js
